[MME] Add Emergency Number List to "Attach Accept" NAS message
Emergency numbers may be configured:

mme:
...
  emergency:
    dnn: internet.sos
    number:
      - digits: 110
        categories: police
      - digits: 112
        categories: [ fire, ambulance ]

The categories are: police, ambulance, fire, marine, mountain

They are sent to the UE during attachment. Some UEs require this list
of numbers. When an emergency call is made, one of the given categories
is indicated towards the IMS in the INVITE message.

Related: SYS#7635
Change-Id: I1b0e2ca0f91863f28216db0deb50c37c99e44a28